🔨 refactor : 파일 정리 - #41
