+++ src/main/java/refresh/acci/domain/analysis/application/usecase/QueryAnalysisUseCase.java
@@ -1,21 +1,40 @@
 package refresh.acci.domain.analysis.application.usecase;
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import refresh.acci.domain.analysis.adapter.in.web.dto.res.AnalysisResultResponse;
+import refresh.acci.domain.analysis.adapter.in.web.dto.res.AnalysisSummaryResponse;
 import refresh.acci.domain.analysis.adapter.out.ai.dto.res.AiResultResponse;
 import refresh.acci.domain.analysis.application.port.out.AnalysisRepositoryPort;
 import refresh.acci.domain.analysis.model.Analysis;
+import refresh.acci.domain.user.model.CustomUserDetails;
+import refresh.acci.global.common.PageResponse;
+import refresh.acci.global.exception.CustomException;
+import refresh.acci.global.exception.ErrorCode;
 
 import java.util.UUID;
 
-@Slf4j
 @Service
 @RequiredArgsConstructor
-public class UpdateAnalysisStatusUseCase {
+@Transactional(readOnly = true)
+public class QueryAnalysisUseCase {
 
     private final AnalysisRepositoryPort analysisRepository;
+
+    public AnalysisResultResponse getAnalysisResult(UUID analysisId, CustomUserDetails userDetails) {
+        Analysis analysis = analysisRepository.getById(analysisId);
+
+        if (userDetails == null || analysis.getUserId() == null || !analysis.getUserId().equals(userDetails.getId())) {
+            throw new CustomException(ErrorCode.ACCESS_DENIED_TO_ANALYSIS);
+        }
+
+        return AnalysisResultResponse.of(analysis);
+    }
+
+    public PageResponse<AnalysisSummaryResponse> getUserAnalyses(Long userId, int page, int size) {
+        return analysisRepository.getUserAnalyses(userId, page, size);
+    }
 
     @Transactional
     public Analysis markProcessing(UUID analysisId, String aiJobId) {
